♻️ Refactor code to validate userId with the jwt

+++ Backend/src/main/java/com/cr1st1an/itemskeeper/backend/controllers/CollectionController.java
@@ -1,11 +1,12 @@
 package com.cr1st1an.itemskeeper.backend.controllers;
 
-import com.cr1st1an.itemskeeper.backend.persistence.entities.User;
-import com.cr1st1an.itemskeeper.backend.persistence.respositories.UserRepository;
 import com.cr1st1an.itemskeeper.backend.services.ICollectionService;
 import com.cr1st1an.itemskeeper.backend.services.IItemService;
 import com.cr1st1an.itemskeeper.backend.services.models.dtos.CollectionDTO;
 import com.cr1st1an.itemskeeper.backend.services.models.dtos.ItemDTO;
+import com.cr1st1an.itemskeeper.backend.services.models.validations.ObjectsValidations;
+import com.cr1st1an.itemskeeper.backend.utils.JWTUtils;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -17,14 +18,16 @@ import java.util.List;
 public class CollectionController {
 
     private final ICollectionService collectionService;
-    private final UserRepository userRepository;
     private final IItemService itemService;
+    private final JWTUtils jwtUtils;
+    private final ObjectsValidations objectsValidations;
 
     @Autowired
-    public CollectionController(ICollectionService collectionService, UserRepository userRepository, IItemService itemService) {
+    public CollectionController(ICollectionService collectionService, IItemService itemService, JWTUtils jwtUtils, ObjectsValidations objectsValidations) {
         this.collectionService = collectionService;
-        this.userRepository = userRepository;
         this.itemService = itemService;
+        this.jwtUtils = jwtUtils;
+        this.objectsValidations = objectsValidations;
     }
 
     @GetMapping("/top")
@@ -33,14 +36,9 @@ public class CollectionController {
         return ResponseEntity.ok(topCollections);
     }
 
-    @PostMapping
-    public ResponseEntity<CollectionDTO> createCollection(@RequestBody CollectionDTO collectionDTO) {
-        User user = userRepository.findById(collectionDTO.getUserId()).orElse(null);
-        if (user == null) {
-            return ResponseEntity.badRequest().build();
-        }
-        CollectionDTO collection = collectionService.createCollection(collectionDTO);
-        return ResponseEntity.ok(collection);
+    @GetMapping("/{collectionId}/items")
+    public ResponseEntity<List<ItemDTO>> getItemsByCollectionId(@PathVariable Long collectionId) {
+        return ResponseEntity.ok(itemService.getItemsByCollectionId(collectionId));
     }
 
     @GetMapping
@@ -50,35 +48,48 @@ public class CollectionController {
 
     @GetMapping("/{collectionId}")
     public ResponseEntity<CollectionDTO> getCollectionById(@PathVariable Long collectionId) {
-        return collectionService.getCollectionById(collectionId)
-                .map(ResponseEntity::ok)
-                .orElseGet(() -> ResponseEntity.notFound().build());
+        return collectionService.getCollectionById(collectionId).map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
     }
 
-    @PutMapping("/{collectionId}")
-    public ResponseEntity<CollectionDTO> updateCollection(
-            @PathVariable Long collectionId,
-            @RequestBody CollectionDTO collectionDTO
-    ) {
-        CollectionDTO updatedCollection = collectionService.updateCollection(collectionId, collectionDTO);
-        return ResponseEntity.ok(updatedCollection);
+    @PostMapping
+    public ResponseEntity<CollectionDTO> createCollection(@RequestBody CollectionDTO collectionDTO, HttpServletRequest request) {
+        try {
+            String token = request.getHeader("Authorization").substring(7);
+            Long userIdFromToken = jwtUtils.getUserIdFromJWT(token);
+            collectionDTO.setUserId(userIdFromToken);
+            CollectionDTO collection = collectionService.createCollection(collectionDTO);
+            return ResponseEntity.ok(collection);
+        } catch (IllegalArgumentException e) {
+            return ResponseEntity.badRequest().body(null);
+        } catch (Exception e) {
+            return ResponseEntity.status(500).body(null);
+        }
+    }
+
+    @PutMapping()
+    public ResponseEntity<?> updateCollection(@RequestBody CollectionDTO collectionDTO, HttpServletRequest request) {
+        try {
+            String token = request.getHeader("Authorization").substring(7);
+            Long userIdFromToken = jwtUtils.getUserIdFromJWT(token);
+            collectionDTO.setUserId(userIdFromToken);
+            CollectionDTO updatedCollection = collectionService.updateCollection(collectionDTO);
+            return ResponseEntity.ok(updatedCollection);
+        } catch (IllegalArgumentException e) {
+            return ResponseEntity.badRequest().body(null);
+        } catch (Exception e) {
+            return ResponseEntity.status(500).body(null);
+        }
     }
 
     @DeleteMapping("/{collectionId}")
-    public ResponseEntity<Void> deleteCollection(@PathVariable Long collectionId) {
+    public ResponseEntity<?> deleteCollection(@PathVariable Long collectionId, HttpServletRequest request) {
+        ResponseEntity<?> validationResponse = objectsValidations.validateCollectionId(request, collectionId);
+        if (validationResponse.getStatusCode().isError()) {
+            return validationResponse;
+        }
+
         collectionService.deleteCollection(collectionId);
         return ResponseEntity.noContent().build();
     }
-
-    @PostMapping("/{collectionId}/items")
-    public ResponseEntity<ItemDTO> createItem(@PathVariable Long collectionId, @RequestBody ItemDTO itemDTO) {
-        itemDTO.setCollectionId(collectionId);
-        ItemDTO createdItem = itemService.createItem(itemDTO);
-        return ResponseEntity.ok(createdItem);
-    }
-
-    @GetMapping("/{collectionId}/items")
-    public ResponseEntity<List<ItemDTO>> getItemsByCollectionId(@PathVariable Long collectionId) {
-        return ResponseEntity.ok(itemService.getItemsByCollectionId(collectionId));
-    }
 }
+
